Added the DAS class within tas.py to remove redundancies
